requirements.txt
annotated-types==0.7.0
anthropic==0.53.0
anyio==4.9.0
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
ecdsa==0.19.1
exceptiongroup==1.3.0
feedparser==6.0.11
flake8==7.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pycodestyle==2.13.0
pydantic==2.11.5
pydantic_core==2.33.2
pyflakes==3.3.2
python-dotenv==1.1.0
python-http-client==3.3.7
requests==2.32.4
sendgrid==6.12.3
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
